+++ src/Properties/Changelog.md
@@ -8,6 +8,8 @@
 - Added the `is_hidden` field to the `Property` model
 - Dropped PHP 8.0 & PHP 8.1 Support
 - Dropped Laravel 9 Support
+- Added PHP 8.3 Support
+- Changed minimum Laravel version to v10.38.2
 
 ## 3.x Series