+++ src/Properties/Changelog.md
@@ -2,8 +2,8 @@
 
 ## 4.x Series
 
-## Unreleased
-##### 2023-XX-YY
+## 4.0.0
+##### 2024-04-25
 
 - Added the `is_hidden` field to the `Property` model
 - Dropped PHP 8.0 & PHP 8.1 Support